SELECTED state for pause menu in UI demo sets highlight color

+++ Library/include/ecs/components.h
@@ -65,7 +65,7 @@ typedef struct {
 
 typedef struct {
   char *text;
-  int color;
+  u16 color;
 } TextComponent;
 
 typedef struct ComponentStorage {
